Move `send_to_paste_service` to services submodule
